+++ resource_planning/models/__init__.py
@@ -1,5 +1,7 @@
-from . import resource_location
-from . import resource
-from . import resource_allocation
-from . import resource_category
 from . import res_config
+from . import res_users
+from . import res_partner
+from . import resource
+from . import resource_location
+from . import resource_category
+from . import resource_allocation
