+++ resource_planning/views/resource_allocation_views.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="resouce_allocation_tree" model="ir.ui.view">
+        <field name="name">resource.allocation.tree</field>
+        <field name="model">resource.allocation</field>
+        <field name="arch" type="xml">
+            <tree string="Resource Allocation" delete="false"
+                  colors="blue:state == 'reservation'; red:state == 'option';grey:state == 'cancel'">
+                <field name="resource_id"/>
+                <field name="serial_number"/>
+                <field name="resource_category_id"/>
+                <field name="location" groups="resource_planning.group_multi_location"/>
+                <field name="partner_id"/>
+                <field name="date_start"/>
+                <field name="date_end"/>
+                <field name="date_lock"/>
+                <field name="state"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="view_resource_allocation_form" model="ir.ui.view">
+        <field name="name">resource.allocation.form</field>
+        <field name="model">resource.allocation</field>
+        <field name="arch" type="xml">
+            <form string="Resource Allocation">
+                <header>
+                    <button name="action_confirm" string="Confirm" type="object" states="option" class="btn-primary"/>
+                    <button name="action_cancel" string="Cancel" type="object" states="booked,option"/>
+                    <button name="action_option" string="Set an option" type="object" states="booked"/>
+                    <field name="state" widget="statusbar"/>
+                </header>
+                <sheet>
+                    <group>
+                        <group>
+                            <field name="resource_id"/>
+                            <field name="resource_category_id"/>
+                            <field name="location" readonly="True" groups="resource_planning.group_multi_location"/>
+                        </group>
+                        <group>
+                            <field name="date_start"/>
+                            <field name="date_end"/>
+                            <field name="date_lock"/>
+                        </group>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="view_resource_allocation_filter" model="ir.ui.view">
+        <field name="name">resource.allocation.search</field>
+        <field name="model">resource.allocation</field>
+        <field name="arch" type="xml">
+            <search string="Search Resource Allocation">
+                <field name="resource_id" string="Resource"/>
+                <field name="partner_id" string="Partner"/>
+                <field name="location" string="location" groups="resource_planning.group_multi_location"/>
+                <filter string="Option" name="option" domain="[('state','=','option')]"/>
+                <filter string="Booked" name="booked" domain="[('state','=','booked')]"/>
+                <filter string="Maintenance" name="maintenance" domain="[('state','=','maintenance')]"/>
+                <separator/>
+                <filter string="Currently" name="currently"
+                        domain="[('date_start','&lt;=', datetime.datetime.combine(context_today(), datetime.time(23, 59, 59))),
+                     ('date_end', '&gt;=', datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)))]"/>
+                <separator/>
+                <group string="Group By...">
+                    <filter string="Resource" domain="[]" context="{'group_by':'resource_id'}"/>
+                    <filter string="Resource Category" domain="[]" context="{'group_by':'resource_category_id'}"/>
+                </group>
+            </search>
+        </field>
+    </record>
+
+    <record id="view_resource_allocation_gantt" model="ir.ui.view">
+        <field name="name">resource.allocation.gantt</field>
+        <field name="model">resource.allocation</field>
+        <field name="type">gantt8</field>
+        <field name="arch" type="xml">
+            <gantt8 editable="False" date_start="date_start" date_stop="date_end" string="Allocations"
+                    default_group_by="serial_number">
+            </gantt8>
+        </field>
+    </record>
+</odoo>
