+++ resource_planning/views/resource_category_views.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="resource_category_tree" model="ir.ui.view">
+        <field name="name">resource.category.tree</field>
+        <field name="model">resource.category</field>
+        <field name="arch" type="xml">
+            <tree string="Resource Category" delete="false">
+                <field name="name"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="view_resource_category_form" model="ir.ui.view">
+        <field name="name">resource.category.form</field>
+        <field name="model">resource.category</field>
+        <field name="arch" type="xml">
+            <form string="Resource Category">
+                <sheet>
+                    <div class="oe_button_box" name="button_box">
+                        <button class="oe_stat_button" type="object" name="toggle_active" icon="fa-archive">
+                            <field name="active" widget="boolean_button" options='{"terminology": "archive"}'/>
+                        </button>
+                    </div>
+                    <group>
+                        <group name="name">
+                            <field name="name"/>
+                            <field name="is_accessory"/>
+                        </group>
+                    </group>
+                    <notebook>
+                        <page name="resources" string="Resources">
+                            <field name="resources">
+                                <tree delete="false" create="false">
+                                    <field name="name"/>
+                                    <field name="code"/>
+                                    <field name="resource_type"/>
+                                    <field name="state"/>
+                                </tree>
+                            </field>
+                        </page>
+                    </notebook>
+                </sheet>
+                <div class="oe_chatter">
+                    <field name="message_follower_ids" widget="mail_followers"/>
+                    <field name="message_ids" widget="mail_thread"/>
+                </div>
+            </form>
+        </field>
+    </record>
+</odoo>
